Refactor equality benchmarks into dedicated modules.
