+++ benchmarks/benchmarks/sperf/__init__.py
@@ -14,7 +14,6 @@ from iris import load_cube
 # TODO: remove uses of PARSE_UGRID_ON_LOAD once UGRID parsing is core behaviour.
 from iris.experimental.ugrid import PARSE_UGRID_ON_LOAD
 
-from .. import TrackAddedMemoryAllocation, on_demand_benchmark
 from ..generate_data.ugrid import make_cubesphere_testfile
 
 
@@ -38,29 +37,3 @@ class FileMixin:
     def load_cube(self):
         with PARSE_UGRID_ON_LOAD.context():
             return load_cube(str(self.file_path))
-
-
-@on_demand_benchmark
-class CubeComparison(FileMixin):
-    """
-    Benchmark time and memory costs of comparing :class:`~iris.cube.Cube`\\ s
-     with attached :class:`~iris.experimental.ugrid.mesh.Mesh`\\ es.
-
-    Uses :class:`FileMixin` as the realistic case will be comparing
-    :class:`~iris.cube.Cube`\\ s that have been loaded from file.
-
-    """
-
-    # Cut down paremt parameters.
-    params = [FileMixin.params[0]]
-
-    def setup(self, c_size, n_levels=1, n_times=1):
-        super().setup(c_size, n_levels, n_times)
-        self.cube = self.load_cube()
-        self.other_cube = self.load_cube()
-
-    def peakmem_eq(self, n_cube):
-        _ = self.cube == self.other_cube
-
-    def time_eq(self, n_cube):
-        _ = self.cube == self.other_cube
